% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2gene.R
\name{h2gene}
\alias{h2gene}
\title{Estimate partitioned gene-level heritability}
\usage{
h2gene(susie_fit, ld, annot, n_sample = 500)
}
\arguments{
\item{susie_fit}{susie fit object which is obtained with
\code{\link[susieR]{susie}} or \code{\link[susieR]{susie_suff_stat}}}

\item{ld}{A \code{n_snp} by \code{n_snp} linkage disequilibrium matrix which
should match the one used to construct \code{susie_fit}}

\item{annot}{A \code{n_snp} by \code{n_annot} \emph{binary} matrix denoting the}

\item{n_sample}{Number of posterior samples to simulate for estimation.
Default to 500.}
}
\value{
A list with the following elements:

\item{hsq}{\code{n_sample} posterior samples of heritability}

\item{ncausal}{\code{n_sample} posterior samples of number of causals}
}
\description{
This function estimate partitioned gene-level heritability
}
