% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2gene.R
\name{h2gene}
\alias{h2gene}
\title{Estimate partitioned gene-level heritability}
\usage{
h2gene(susie_fit, ld, annot, n_sample = 500)
}
\arguments{
\item{susie_fit}{A \code{n_indiv} by \code{n_snp} genotype matrix}

\item{ld}{A \code{n_snp} by \code{n_snp} linkage disequilibrium matrix}

\item{annot}{Number of individuals \code{n_indiv} for the LD matrix}

\item{n_sample}{Heritability explained by the genotype
\eqn{\frac{\text{Var}[X\beta]}{\text{Var}[y]}}}

\item{beta}{\code{n_snp} by \code{n_sim} simulated effect sizes}
}
\value{
A list with the following elements:

\item{beta_hat}{Simulated marginal effects
  \code{beta_hat[, i]} corresponds to ith simulation}

\item{e}{Simulated environmental noise}
}
\description{
This function estimate partitioned gene-level heritability
}
