% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_gwas}
\alias{simulate_gwas}
\title{Simulate GWAS from either individual level genotype or LD matrix}
\usage{
simulate_gwas(ld, n_indiv, hsq, beta)
}
\arguments{
\item{ld}{A \code{n_snp} by \code{n_snp} linkage disequilibrium matrix}

\item{n_indiv}{Number of individuals \code{n_indiv} for the LD matrix}

\item{hsq}{Heritability explained by the genotype
\eqn{\frac{\text{Var}[X\beta]}{\text{Var}[y]}}}

\item{beta}{\code{n_snp} by \code{n_sim} simulated effect sizes}

\item{X}{A \code{n_indiv} by \code{n_snp} genotype matrix}
}
\value{
A list with the following elements:

\item{beta_hat}{Simulated marginal effects
  \code{beta_hat[, i]} corresponds to ith simulation}

\item{e}{Simulated environmental noise}
}
\description{
This function simulate GWAS marginal standardized beta
\eqn{\hat{\beta} = \frac{X^T y}{N}} with genotype \eqn{X} and phenotype
simulated under the standard linear model \eqn{y = Xb + e}. It supports
two modes: (1) simulation from the individual level genotype X. (2)
simulation from the LD matrix \eqn{X^T X}. The second method is not so
straightforward and warrant some explanations:
\deqn{\hat{\beta} = \frac{X^T y}{N} = \frac{X^T (X b + e)}{N} =
\frac{X^T X}{N}b + \frac{X^T e}{N}}
Therefore, we can simulate \eqn{\hat{\beta}} with a multivariate normal
distribution with mean \eqn{\frac{X^T X}{N}b} and variance
\eqn{\frac{1}{N} \frac{X^T X}{N} \sigma_e^2}
}
